+++ src/app/app.module.ts
@@ -9,7 +9,8 @@ import { WelcomeComponent } from './pages/welcome/welcome.component';
 import { ProductListComponent } from './components/product-list/product-list.component';
 import { HeaderComponent } from './components/header/header.component';
 import { FilterByTagComponent } from './components/filter-by-tag/filter-by-tag.component';
-import { CentimesToEurosPipe, ProductCardsComponent } from './components/product-cards/product-cards.component';
+import { ProductCardsComponent } from './components/product-cards/product-cards.component';
+import { CentimesToEurosPipe } from './components/product-cards/product-cards.component';
