+++ src/app/components/filter-by-tag/filter-by-tag.component.html
@@ -1,7 +1,13 @@
 <ul class="filterTag">
-  <li><a routerLink="">Base tomate</a></li>
-  <li><a routerLink="">Base blanche</a></li>
-  <li><a routerLink="">Veggie</a></li>
-  <li><a routerLink="">Desserts</a></li>
-  <li><a routerLink="">Boissons</a></li>
+  <li><a routerLink="/all-products" (click)="greet(products)" >
+    >Base tomate </a> </li>
+  <li><a routerLink="/all-products" (click)="('white')"
+    > Base blanche </a></li>
+  <li><a routerLink="/all-products" (click)="('veggie')"
+    >Veggie</a></li>
+  <li><a routerLink="/all-products" (click)="('dessert')"
+    >Desserts</a></li>
+  <li><a routerLink="/all-products" (click)="('drink')"
+    >Boissons</a></li>
 </ul>
+
