+++ src/app/components/product-cards/product-cards.component.css
@@ -1,26 +1,38 @@
 /* GENERAL  */
-
 li {
     list-style-type: none;
+    background-color: white;
+    padding: 10px;
 }
+h2 {
+    font-size: 20px;
+    font-family: 'Playfair Display', serif;
+    font-style: italic;
+    margin-top: 54px;
+}
+
 img {
     width: 207px;
 }
+
 h3 {
     font-size: 18px;
     font-family: 'Playfair Display', serif;
-
 }
-h3, p {
+
+h3,
+p {
     margin: 0;
 }
+
 p {
     font-size: 16px;
     font-family: 'Inter', sans-serif;
-
 }
-li {
-    background-color: white;
-    padding: 10px;
-    
+/* DISPLAY FLEX  */
+ul {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 30px;
+    padding: 0;
 }
