+++ src/app/components/product-list/product-list.component.css
@@ -1,24 +1,8 @@
-
-
 h2 {
+    text-align: center;
     font-size: 20px;
     font-family: 'Playfair Display', serif;
-    font-style:  italic;
-}
-
-
-
-
-/* display-flex  */
-.column {
-    display: flex;
-    flex-direction: column;
-    flex-wrap: wrap;
-    margin: 0 57px;
-}
-.flexWrap {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 30px;
-    padding: 0;
+    font-style: italic;
+    margin-top: 54px;
+    color: #E12417;
 }
