+++ src/app/components/product-list/product-list.component.html
@@ -1,34 +1,5 @@
- <main class="column"> 
-
-    <div>
-    <h2>Les Pizzas della Mamma</h2>
-
-    <ul class="flexWrap">  
-        <app-product-cards *ngFor="let product of newData['pizza']" [product] = "product"> </app-product-cards>
-
-        </ul>
-    </div>
-
-    <div>
-        <h2>Les Desserts della Mamma</h2>
-
-    <ul class="flexWrap">  
-        <app-product-cards *ngFor="let product of newData['dessert']" [product] = "product"> </app-product-cards>
-
-        </ul>
-    </div>
-
-    <div>
-        <h2>Les Boissons della Mamma</h2>
-
-
-    <ul class="flexWrap">  
-        <app-product-cards *ngFor="let product of newData['drink']" [product] = "product"> </app-product-cards>
-
-        </ul>
-    </div>
-
-<!-- note :  on accède aux différentes catégories dans l'objet MyType en utilisant leur nom en tant que clé, par exemple newData['pizza'] pour accéder à la catégorie "pizza". -->
-    </main>
-        
-
+<ul>
+    <app-product-cards *ngFor="let product of products" [productCards]="product">
+    </app-product-cards>
+</ul>
+<h2> Buon Appetito !</h2>
