+++ src/app/pages/all-products/all-products.component.html
@@ -1,2 +1,2 @@
-<app-filter-by-tag> </app-filter-by-tag>
-<app-product-list></app-product-list>
+<app-filter-by-tag > </app-filter-by-tag>
+<app-product-list [products]="productsCategory"> </app-product-list>
